Add a command abstraction
In an effort to make it so that different Bazel labels will have different
behavior when running iBazel we added an abstraction of what command is
running. This abstraction encapsulates how the command should start, terminate
itself and respond to changes in build dependencies. Different commands will
respond to changes in different ways. By default, the command will terminate
and restart. In the future we envision some Bazel targets being having a label
which tells iBazel to not kill the command when changes are detected. Instead
they can do different sorts of behavior (ie performing a hotswap of resources
for a java binary).