Add simple e2e test
This e2e test runs iBazel and creates a file system for it to watch.
Then it runs a program inside the FS. Recording the PID and the
process's output, it then modifies the mock source files and BUILD files
and verifies that iBazel correctly restarts the subprocess each time.

This is also a good base for testing more elaborate problems like the
perpernicious vim save bug where the watchers are removed. It will also
be useful when testing the notifying launcher. It will also be useful
for avoiding functionality regressions on systems I don't often use (see
Darwin/Windows).