Update to latest rules_go
Also replace bazel_integraion_tests repo with the golang native bazel
testing suite. This doesn't provide coverage for old versions of Bazel,
but it is less of a maintanence burdon since it uses system bazel.

If testing on old versions is deemed high enough value, we can change
the CI config to run this against `current_release` and
`current_release - n` versions of Bazel.